캘린더 리팩토링 validation 추가 참여자 리스트 쿼리 수정 참여자수에 따른 모임 상태 변경로직 추가

+++ src/main/java/com/gmail/dlwk0807/dagachi/service/GroupAttendService.java
@@ -7,6 +7,7 @@ import com.gmail.dlwk0807.dagachi.dto.group.GroupResponseDTO;
 import com.gmail.dlwk0807.dagachi.dto.member.MemberAttendResponseDTO;
 import com.gmail.dlwk0807.dagachi.entity.Group;
 import com.gmail.dlwk0807.dagachi.entity.GroupAttend;
+import com.gmail.dlwk0807.dagachi.entity.GroupStatus;
 import com.gmail.dlwk0807.dagachi.entity.Member;
 import com.gmail.dlwk0807.dagachi.repository.GroupAttendRepository;
 import com.gmail.dlwk0807.dagachi.repository.GroupRepository;
@@ -17,6 +18,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 import static com.gmail.dlwk0807.dagachi.util.SecurityUtil.getCurrentMemberId;
@@ -43,11 +45,24 @@ public class GroupAttendService {
         }
 
         Group group = groupRepository.findById(groupId).orElseThrow(() -> new CustomRespBodyException("모임정보가 없습니다."));
+
+        //모임 최대인원 확인
+        Long groupMaxNum = group.getGroupMaxNum();
+        int attendNum = group.getGroupAttendList().size();
+        if (groupMaxNum.compareTo((long) attendNum) <= 0) {
+            throw new CustomRespBodyException("인원수가 모두 찼습니다.");
+        }
+        if (groupMaxNum.equals((long) (attendNum + 1))) {
+            group.updateStatus(GroupStatus.FULL);
+        }
+
         GroupAttend groupAttend = GroupAttend.builder()
                 .group(group)
                 .member(member)
                 .build();
         groupAttendRepository.save(groupAttend);
+
+        group.getGroupAttendList().add(groupAttend);
 
         return GroupResponseDTO.of(group);
     }
@@ -67,8 +82,20 @@ public class GroupAttendService {
             throw new CustomRespBodyException("참여하지 않은 모임입니다.");
         }
 
-        GroupAttend groupAttend = groupAttendRepository.findByGroupIdAndMemberId(groupId, currentMemberId).orElseThrow();
+        Group group = groupRepository.findById(groupId).orElseThrow(() -> new CustomRespBodyException("존재하지 않는 모임입니다."));
+        if (currentMemberId.equals(group.getMemberId())) {
+            throw new CustomRespBodyException("모임장은 참여 취소를 할 수 없습니다.");
+        }
+
+        GroupAttend groupAttend = groupAttendRepository.findByGroupIdAndMemberId(groupId, currentMemberId).orElseThrow(() -> new CustomRespBodyException("참여하지 않은 모임입니다."));
         groupAttendRepository.delete(groupAttend);
+
+        //모임 상태 변경
+        if (GroupStatus.FULL.equals(group.getStatus())) {
+            group.updateStatus(GroupStatus.WAITING);
+        }
+
+        group.getGroupAttendList().remove(groupAttend);
 
         return "ok";
     }
